refactor:make stub package
