+++ src/test/java/ladder/model/LadderTest.java
@@ -1,11 +1,10 @@
 package ladder.model;
 
-import org.junit.jupiter.api.Test;
+import ladder.model.stub.LadderMovingStub;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
