+++ src/test/java/ladder/model/stub/LadderMovingStub.java
@@ -1,4 +1,6 @@
-package ladder.model;
+package ladder.model.stub;
+
+import ladder.model.MoveStrategy;
 
 import java.util.ArrayList;
 import java.util.Arrays;
